Add support for specifying a node bootstrap file while creating and updating a cluster.
It adds a --node-bootstrap-file option to the 'cluster create' and 'cluster update' commands.
In the sdk, I have added this option to the constructor for ClusterInfo. I think we need to
restructure where to specify these settings. Waiting for the restructure to happen at the API
level before doing that.