Add support in the SDK for composite commands.
The CLI doesn't support them yet. I don't know of a good way of doing it.
If you consturct the sub_commands using the parse method of the corresponding
classes, there would be extra 'can_notify' and 'label' keys inside them
which the API would ignore. Only the options given directly to the composite
commands are acted upon by the API.