fix: usr: Support different qubole version for cluster methods. Pull Request #144.
When creating clusters with the v1.3 of the API we still use the v1.2 endpoint
during cluster creation. The _parse_create_update method accepts the
api-version only to distinguish the request parameters. But the create, update
and clone methods still use the default configuration of Qubole.agent(..) which
is set to v1.2 unless all the actions in the session are configure(..)`ed with
the v1.3 of the API.

If you want to perform some of the unsupported actions with v1.3 of the
cluster API but fall back to v1.2 for commands APIs we explicitly switch the
version in the create(..)/update(..)/clone(..) calls which thereby alter the
version for that API call alone and leave the cached_agent as is for the v1.2
APIs.