fix: usr: Bug fix for access denied error while downloading file. Pull Request #190.
SDK-191: Aws throws Access denied errors in situations where the versioning on
s3 files is enabled and an attempt to download the file is done by specifying the
version(default beahvior by aws boto2). This is seen in boto2 and the workaround is to
retry downloading again without specifying the version of the file.