fix comparisons to literals
Change comparisons to literals from `is` to `==`.

Comparisons to literals with `is` aren't Pythonic because `is` uses the underlying object ID to determine whether objects match, whereas `==` does a value comparison. By switching the comparisons to `==`, we have a more consistent set of comparisons, because we don't rely on object IDs but rather the values themselves, which is what we're looking for, anyhow.